refa: Controller 역할 분리
