+++ src/main/java/controller/game/Command.java
@@ -1,4 +1,4 @@
-package controller;
+package controller.game;
 
 import java.util.Arrays;
