+++ src/main/java/controller/room/GameRoomAction.java
@@ -1,8 +1,8 @@
-package controller;
+package controller.room;
 
 import java.util.List;
 
 @FunctionalInterface
-public interface Action {
+public interface GameRoomAction {
     void execute(List<String> inputs);
 }
