+++ src/test/java/controller/CommandTest.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
+import controller.game.Command;
+
 class CommandTest {
 
     @ParameterizedTest
